move database actions into data-access
