+++ libs/data-access/src/index.ts
@@ -7,6 +7,8 @@ export * from './lib/folio';
 export * from './lib/glossary';
 export * from './lib/imprint';
 export * from './lib/import-job';
+export * from './lib/import-preview';
+export * from './lib/import-storage';
 export * from './lib/panel-url';
 export * from './lib/library';
 export * from './lib/passage';
