+++ libs/lib-import/src/lib/index.ts
@@ -1,4 +1,3 @@
 export * from './docx';
 export * from './import-job';
 export * from './preview';
-export * from './types';
