+++ libs/lib-import/src/lib/preview.ts
@@ -7,7 +7,7 @@ import type {
   NormalizedParagraph,
   PassageInsertOperation,
   PreviewAnnotationOperation,
-} from './types';
+} from '@eightyfourthousand/data-access';
 
 /**
  * Mapping configuration for a recognized import section heading.
